Minor Bug Fixes and Modified Database
Converted price to retail price and added wholesale price with what we have working currently, I also migrated the old Console testing to a new file and updated it to work with the new database format. Also made database values viewable in the modify menu. (Cory you're going to need to add a column for wholesale price)

Also fixed some bugs in the code when interacting with the database, such as the name being the barcode.